feat(api/user): move delete_user_from_db to database module
