+++ api/src/routes/user/user/delete.rs
@@ -4,12 +4,12 @@ use actix_web::{
 };
 use anyhow::Context;
 use avina_wire::user::User;
-use sqlx::{Executor, MySql, MySqlPool, Transaction};
+use sqlx::MySqlPool;
 
 use super::UserIdParam;
 use crate::{
     authorization::require_admin_user,
-    error::{MinimalApiError, NormalApiError},
+    database::user::user::delete_user_from_db, error::NormalApiError,
 };
 
 #[tracing::instrument(name = "user_delete")]
@@ -29,29 +29,4 @@ pub async fn user_delete(
         .await
         .context("Failed to commit transaction")?;
     Ok(HttpResponse::NoContent().finish())
-}
-
-#[tracing::instrument(name = "delete_user_from_db", skip(transaction))]
-pub async fn delete_user_from_db(
-    transaction: &mut Transaction<'_, MySql>,
-    user_id: u64,
-) -> Result<(), MinimalApiError> {
-    let query = sqlx::query!(
-        r#"
-        DELETE IGNORE FROM user_user
-        WHERE id = ?
-        "#,
-        user_id
-    );
-    let result = transaction
-        .execute(query)
-        .await
-        .context("Failed to execute delete query")?;
-    if result.rows_affected() == 0 {
-        return Err(MinimalApiError::ValidationError(
-            // TODO: test that this message is really correct
-            "Failed to delete user, either it doesn't exist or other data still depends on it.".to_string(),
-        ));
-    }
-    Ok(())
 }
